+++ pypesto/optimize/optimizer.py
@@ -4,12 +4,13 @@ import re
 import abc
 import time
 import logging
-from typing import Dict
+from typing import Dict, Optional
 
 from ..objective import (OptimizerHistory, HistoryOptions, CsvHistory)
 from ..objective.history import HistoryBase
 from ..problem import Problem
 from .result import OptimizerResult
+from ..objective.constants import MODE_FUN, FVAL, GRAD
 
 try:
     import ipopt
@@ -30,6 +31,16 @@ try:
     import cma
 except ImportError:
     cma = None
+
+try:
+    import nlopt
+except ImportError:
+    nlopt = None
+
+try:
+    import fides
+except ImportError:
+    fides = None
 
 EXITFLAG_LOADED_FROM_FILE = -99
 
@@ -218,7 +229,6 @@ def read_result_from_file(problem: Problem, history_options: HistoryOptions,
 class Optimizer(abc.ABC):
     """
     This is the optimizer base class, not functional on its own.
-
     An optimizer takes a problem, and possibly a start point, and then
     performs an optimization. It returns an OptimizerResult.
     """
@@ -241,7 +251,6 @@ class Optimizer(abc.ABC):
     ) -> OptimizerResult:
         """"
         Perform optimization.
-
         Parameters
         ----------
         problem:
@@ -263,6 +272,13 @@ class Optimizer(abc.ABC):
         Create default options specific for the optimizer.
         """
         return None
+
+
+def check_finite_bounds(lb, ub):
+    """Raise if bounds are not finite."""
+    if not np.isfinite(lb).all() or not np.isfinite(ub).all():
+        raise ValueError('Selected optimizer cannot work with unconstrained '
+                         'optimization problems.')
 
 
 class ScipyOptimizer(Optimizer):
@@ -439,7 +455,8 @@ class IpoptOptimizer(Optimizer):
 
         if ipopt is None:
             raise ImportError(
-                "This optimizer requires an installation of ipopt."
+                "This optimizer requires an installation of ipopt. You can "
+                "install ipopt via `pip install ipopt`."
             )
 
         objective = problem.objective
@@ -481,6 +498,9 @@ class DlibOptimizer(Optimizer):
         self.options = options
         if self.options is None:
             self.options = DlibOptimizer.get_default_options(self)
+        elif 'maxiter' not in self.options:
+            raise KeyError('Dlib options are missing the key word '
+                           'maxiter.')
 
     @fix_decorator
     @time_decorator
@@ -495,16 +515,18 @@ class DlibOptimizer(Optimizer):
 
         lb = problem.lb
         ub = problem.ub
+        check_finite_bounds(lb, ub)
         objective = problem.objective
 
         if dlib is None:
             raise ImportError(
-                "This optimizer requires an installation of dlib."
+                "This optimizer requires an installation of dlib. You can "
+                "install dlib via `pip install dlib`."
             )
 
         if not objective.has_fun:
-            raise Exception("For this optimizer, the objective must "
-                            "be able to return function values.")
+            raise ValueError("For this optimizer, the objective must "
+                             "be able to return function values.")
 
         # dlib requires variable length arguments
         def get_fval_vararg(*x):
@@ -526,7 +548,7 @@ class DlibOptimizer(Optimizer):
         return False
 
     def get_default_options(self):
-        return {}
+        return {'maxiter': 10000}
 
 
 class PyswarmOptimizer(Optimizer):
@@ -555,7 +577,11 @@ class PyswarmOptimizer(Optimizer):
         ub = problem.ub
         if pyswarm is None:
             raise ImportError(
-                "This optimizer requires an installation of pyswarm.")
+                "This optimizer requires an installation of pyswarm.You can "
+                "install pyswarm via `pip install pyswarm."
+            )
+
+        check_finite_bounds(lb, ub)
 
         xopt, fopt = pyswarm.pso(
             problem.objective.get_fval, lb, ub, **self.options)
@@ -609,12 +635,17 @@ class CmaesOptimizer(Optimizer):
 
         lb = problem.lb
         ub = problem.ub
+
+        check_finite_bounds(lb, ub)
+
         sigma0 = self.par_sigma0 * np.median(ub - lb)
         self.options['bounds'] = [lb, ub]
 
         if cma is None:
             raise ImportError(
-                "This optimizer requires an installation of cma.")
+                "This optimizer requires an installation of cma. You can "
+                "install cma via `pip install cma."
+            )
 
         result = cma.CMAEvolutionStrategy(
             x0, sigma0, inopts=self.options,
@@ -622,6 +653,247 @@ class CmaesOptimizer(Optimizer):
 
         optimizer_result = OptimizerResult(x=np.array(result[0]),
                                            fval=result[1])
+
+        return optimizer_result
+
+    def is_least_squares(self):
+        return False
+
+
+class NLoptOptimizer(Optimizer):
+    """
+    Global/Local optimization using NLopt.
+    Package homepage: https://nlopt.readthedocs.io/en/latest/
+    """
+
+    def __init__(self, method=None, local_method=None, options: Dict = None,
+                 local_options: Dict = None):
+        """
+        Parameters
+        ----------
+        method:
+            Local or global Optimizer to use for minimization.
+        local_method:
+            Local method to use in combination with the global optimizer (
+            for the MLSL family of solvers) or to solve a subproblem (for the
+            AUGLAG family of solvers)
+        options:
+            Optimizer options. scipy option `maxiter` is automatically
+            transformed into `maxeval` and takes precedence.
+        local_options:
+            Optimizer options for the local method
+        """
+
+        super().__init__()
+
+        if options is None:
+            options = {}
+        elif 'maxiter' in options:
+            options['maxeval'] = options.pop('maxiter')
+        if local_options is None:
+            local_options = {}
+        self.options = options
+        self.local_options = local_options
+
+        if nlopt is None:
+            raise ImportError(
+                "This optimizer requires an installation of NLopt. You can "
+                "install NLopt via `pip install nlopt`.")
+
+        if method is None:
+            method = nlopt.LD_LBFGS
+
+        needs_local_method = [
+            nlopt.G_MLSL, nlopt.G_MLSL_LDS, nlopt.GD_MLSL, nlopt.GD_MLSL_LDS,
+            nlopt.AUGLAG, nlopt.AUGLAG_EQ
+        ]
+
+        if local_method is None and method in needs_local_method:
+            local_method = nlopt.LD_LBFGS
+
+        if local_method is not None and method not in needs_local_method:
+            raise ValueError(f'Method "{method}" does not allow a local '
+                             f'method. Please set `local_method` to None.')
+
+        self.local_methods = [
+            nlopt.LD_VAR1, nlopt.LD_VAR2, nlopt.LD_TNEWTON_PRECOND_RESTART,
+            nlopt.LD_TNEWTON_PRECOND, nlopt.LD_TNEWTON_RESTART,
+            nlopt.LD_TNEWTON, nlopt.LD_LBFGS,
+            nlopt.LD_SLSQP, nlopt.LD_CCSAQ, nlopt.LD_MMA, nlopt.LN_SBPLX,
+            nlopt.LN_NELDERMEAD, nlopt.LN_PRAXIS, nlopt.LN_NEWUOA,
+            nlopt.LN_NEWUOA_BOUND, nlopt.LN_BOBYQA, nlopt.LN_COBYLA,
+        ]
+        self.global_methods = [
+            nlopt.GN_ESCH, nlopt.GN_ISRES, nlopt.GN_AGS, nlopt.GD_STOGO,
+            nlopt.GD_STOGO_RAND, nlopt.G_MLSL, nlopt.G_MLSL_LDS, nlopt.GD_MLSL,
+            nlopt.GD_MLSL_LDS, nlopt.GN_CRS2_LM, nlopt.GN_ORIG_DIRECT,
+            nlopt.GN_ORIG_DIRECT_L, nlopt.GN_DIRECT, nlopt.GN_DIRECT_L,
+            nlopt.GN_DIRECT_L_NOSCAL, nlopt.GN_DIRECT_L_RAND,
+            nlopt.GN_DIRECT_L_RAND_NOSCAL,
+        ]
+        self.hybrid_methods = [
+            nlopt.AUGLAG, nlopt.AUGLAG_EQ
+        ]
+        methods = self.local_methods + self.global_methods + \
+            self.hybrid_methods
+
+        if method not in methods:
+            raise ValueError(f'"{method}" is not a valid method. Valid '
+                             f'methods are: {methods}')
+
+        self.method = method
+
+        if local_method is not None and local_method not in self.local_methods:
+            raise ValueError(f'"{local_method}" is not a valid method. Valid '
+                             f'methods are: {self.local_methods}')
+
+        self.local_method = local_method
+
+    @fix_decorator
+    @time_decorator
+    @history_decorator
+    def minimize(
+            self,
+            problem: Problem,
+            x0: np.ndarray,
+            id: str,
+            history_options: HistoryOptions = None,
+    ) -> OptimizerResult:
+
+        opt = nlopt.opt(self.method, problem.dim)
+
+        valid_options = ['ftol_abs', 'ftol_rel', 'xtol_abs', 'xtol_rel',
+                         'stopval', 'x_weights', 'maxeval', 'maxtime',
+                         'initial_step']
+
+        def set_options(o, options):
+            for option, value in options.items():
+                if option not in valid_options:
+                    raise ValueError(
+                        f'"{option}" is not a valid option. Valid '
+                        f'options are: {valid_options}')
+                getattr(o, f'set_{option}')(value)
+
+        if self.local_method is not None:
+            local_opt = nlopt.opt(self.local_method, problem.dim)
+            set_options(local_opt, self.local_options)
+            opt.set_local_optimizer(local_opt)
+
+        if self.method in self.global_methods:
+            check_finite_bounds(problem.ub, problem.lb)
+
+        opt.set_lower_bounds(problem.lb)
+        opt.set_upper_bounds(problem.ub)
+
+        def nlopt_objective(x, grad):
+            if grad.size > 0:
+                sensi_orders = (0, 1)
+            else:
+                sensi_orders = (0,)
+            r = problem.objective(x, sensi_orders, MODE_FUN, True)
+            if grad.size > 0:
+                grad[:] = r[GRAD]  # note that this must be inplace
+            return r[FVAL]
+
+        opt.set_min_objective(nlopt_objective)
+
+        set_options(opt, self.options)
+        try:
+            result = opt.optimize(x0)
+            msg = 'Finished Successfully.'
+        except (nlopt.RoundoffLimited, nlopt.ForcedStop, ValueError,
+                RuntimeError, MemoryError) as e:
+            result = None
+            msg = str(e)
+        optimizer_result = OptimizerResult(x=result,
+                                           fval=opt.last_optimum_value(),
+                                           message=msg,
+                                           exitflag=opt.last_optimize_result())
+
+        return optimizer_result
+
+    def is_least_squares(self):
+        return False
+
+
+class FidesOptimizer(Optimizer):
+    """
+    Global/Local optimization using the trust region optimizer fides.
+    Package Homepage: https://fides-optimizer.readthedocs.io/en/latest
+    """
+
+    def __init__(
+            self,
+            hessian_update: Optional['fides.HessianApproximation'] = None,
+            options: Optional[Dict] = None,
+            verbose: Optional[int] = logging.INFO
+    ):
+        """
+        Parameters
+        ----------
+        options:
+            Optimizer options.
+        hessian_update:
+            Hessian update strategy. If this is None, Hessian (approximation)
+            computed by problem.objective will be used (default).
+        """
+
+        super().__init__()
+
+        if hessian_update is not None and \
+                not isinstance(hessian_update, fides.HessianApproximation):
+            raise ValueError('Incompatible type for hessian update, '
+                             'must be fides.HessianApproximation, '
+                             f'was {type(hessian_update)}.')
+        if options is None:
+            options = {}
+
+        self.verbose = verbose
+        self.options = options
+        self.hessian_update = hessian_update
+
+    @fix_decorator
+    @time_decorator
+    @history_decorator
+    def minimize(
+            self,
+            problem: Problem,
+            x0: np.ndarray,
+            id: str,
+            history_options: HistoryOptions = None,
+    ) -> OptimizerResult:
+
+        if fides is None:
+            raise ImportError(
+                "This optimizer requires an installation of fides. You can "
+                "install fides via `pip install fides`."
+            )
+
+        args = {'mode': MODE_FUN}
+        if self.hessian_update is None:
+            args['sensi_orders'] = (0, 1, 2)
+        else:
+            args['sensi_orders'] = (0, 1)
+
+        opt = fides.Optimizer(
+            fun=problem.objective, funargs=args, ub=problem.ub, lb=problem.lb,
+            verbose=self.verbose,
+            hessian_update=self.hessian_update, options=self.options
+        )
+
+        try:
+            opt.minimize(x0)
+            if opt.converged:
+                msg = 'Finished Successfully.'
+            else:
+                msg = 'Failed to converge'
+        except RuntimeError as err:
+            msg = str(err)
+
+        optimizer_result = OptimizerResult(
+            x=opt.x_min, fval=opt.fval_min, grad=opt.grad_min, hess=opt.hess,
+            message=msg, exitflag=opt.exitflag
+        )
 
         return optimizer_result