+++ test/profile/test_profile.py
@@ -4,7 +4,6 @@ This is for testing profiling of the pypesto.Objective.
 
 import numpy as np
 import unittest
-import test.test_objective as test_objective
 from copy import deepcopy
 import warnings
 
@@ -13,14 +12,17 @@ import pypesto.optimize as optimize
 import pypesto.profile as profile
 import pypesto.visualize as visualize
 from pypesto import ObjectiveBase
-from .visualize import close_fig
+
+from ..visualize import close_fig
+from ..util import rosen_for_sensi
+from numpy.testing import assert_almost_equal
 
 
 class ProfilerTest(unittest.TestCase):
 
     @classmethod
     def setUp(cls):
-        cls.objective: ObjectiveBase = test_objective.rosen_for_sensi(
+        cls.objective: ObjectiveBase = rosen_for_sensi(
             max_sensi_order=2, integrated=True
         )['obj']
 
@@ -70,6 +72,33 @@ class ProfilerTest(unittest.TestCase):
             # standard plotting
             visualize.profiles(result, profile_list_ids=i_run)
             visualize.profile_cis(result, profile_list=i_run)
+
+    def test_engine_profiling(self):
+        # loop over all possible engines
+        # engine=None will be used for comparison
+        engines = [None,
+                   pypesto.engine.SingleCoreEngine(),
+                   pypesto.engine.MultiProcessEngine(),
+                   pypesto.engine.MultiThreadEngine()
+                   ]
+        for engine in engines:
+            # run profiling, profile results get appended
+            # in self.result.profile_result
+            profile.parameter_profile(
+                problem=self.problem,
+                result=self.result,
+                optimizer=self.optimizer,
+                next_guess_method='fixed_step',
+                engine=engine,)
+
+        # check results
+        for count, _engine in enumerate(engines[1:]):
+            for j in range(len(self.result.profile_result.list[0])):
+                assert_almost_equal(
+                    self.result.profile_result.list[0][j]['x_path'],
+                    self.result.profile_result.list[count][j]['x_path'],
+                    err_msg='The values of the profiles for'
+                            ' the different engines do not match')
 
     def test_selected_profiling(self):
         # create options in order to ensure a short computation time
@@ -175,8 +204,8 @@ class ProfilerTest(unittest.TestCase):
 
 # dont make this a class method such that we dont optimize twice
 def test_profile_with_history():
-    objective = test_objective.rosen_for_sensi(max_sensi_order=2,
-                                               integrated=False)['obj']
+    objective = rosen_for_sensi(max_sensi_order=2,
+                                integrated=False)['obj']
 
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
@@ -204,7 +233,7 @@ def test_profile_with_history():
 @close_fig
 def test_profile_with_fixed_parameters():
     """Test using profiles with fixed parameters."""
-    obj = test_objective.rosen_for_sensi(max_sensi_order=1)['obj']
+    obj = rosen_for_sensi(max_sensi_order=1)['obj']
 
     lb = -2 * np.ones(5)
     ub = 2 * np.ones(5)